Added a build script for quick refresh, and created 'example' folder for various sample executions.
Also found a bug in mga_1dsm problem, which messed up the optimisation, if the optimal flyby ratio is > 100. Adjusted, so that upper bound is the planet SOI, and lower bound is the safe_radius.

Also did a tweak in the parameter hashing to store apart MGA and MGA-1DSM